refactor: split players_manager.rs into functional modules
Break the single 2259-line file into server/players_manager/{mod,
accessors, targeting, consumables, turn_flow, game_atk_effect}.rs,
grouped by concern (character accessors, targeting logic, consumable
usage, turn/round flow, GameAtkEffect log formatting). Each module
carries its own unit tests instead of one monolithic test block.
No behavior change — same public API, same 348 passing tests.
